fix(core): respect --existing flag when branch exists locally and on remote
Fixes #4

When using `gren create --existing <branch>` for a branch that exists
both locally and on remote (in sync or behind), the code incorrectly
used the `-b` flag which tried to create a new branch, causing
"branch already exists" error.

Now checks `IsNewBranch` flag in this case:
- If true: recreate branch from remote (existing behavior)
- If false (--existing): use existing local branch directly

Also split worktree_test.go into logical parts for better maintainability:
- worktree_test.go: basic operations (~700 lines)
- worktree_create_test.go: create variations + regression test
- worktree_stale_test.go: stale detection tests
- worktree_submodule_test.go: submodule detection tests
