+++ internal/core/worktree_test.go
@@ -2,7 +2,6 @@ package core
 
 import (
 	"context"
-	"fmt"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -504,77 +503,6 @@ func TestCopyDir(t *testing.T) {
 	})
 }
 
-func TestCreateWorktreeWithCustomDir(t *testing.T) {
-	dir, manager, cleanup := setupTestEnvironment(t)
-	defer cleanup()
-
-	ctx := context.Background()
-
-	t.Run("create worktree in custom directory", func(t *testing.T) {
-		customDir, err := os.MkdirTemp("", "gren-custom-worktrees-*")
-		if err != nil {
-			t.Fatalf("failed to create custom dir: %v", err)
-		}
-		defer os.RemoveAll(customDir)
-
-		req := CreateWorktreeRequest{
-			Name:        "custom-location",
-			IsNewBranch: true,
-			WorktreeDir: customDir,
-		}
-
-		_, err = manager.CreateWorktree(ctx, req)
-		if err != nil {
-			t.Fatalf("CreateWorktree() error: %v", err)
-		}
-
-		// Verify worktree was created in custom location
-		worktreePath := filepath.Join(customDir, "custom-location")
-		if _, err := os.Stat(worktreePath); os.IsNotExist(err) {
-			t.Error("worktree was not created in custom directory")
-		}
-	})
-
-	_ = dir // keep reference for cleanup
-}
-
-func TestCreateWorktreeWithBaseBranch(t *testing.T) {
-	dir, manager, cleanup := setupTestEnvironment(t)
-	defer cleanup()
-
-	ctx := context.Background()
-
-	t.Run("create worktree from specific base branch", func(t *testing.T) {
-		// Get current branch name
-		out, _ := exec.Command("git", "-C", dir, "rev-parse", "--abbrev-ref", "HEAD").Output()
-		baseBranch := string(out[:len(out)-1])
-
-		req := CreateWorktreeRequest{
-			Name:        "from-base",
-			BaseBranch:  baseBranch,
-			IsNewBranch: true,
-		}
-
-		_, err := manager.CreateWorktree(ctx, req)
-		if err != nil {
-			t.Fatalf("CreateWorktree() error: %v", err)
-		}
-
-		// Verify worktree was created
-		worktrees, _ := manager.ListWorktrees(ctx)
-		found := false
-		for _, wt := range worktrees {
-			if wt.Name == "from-base" {
-				found = true
-				break
-			}
-		}
-		if !found {
-			t.Error("worktree 'from-base' not found after creation")
-		}
-	})
-}
-
 func TestListWorktreesEdgeCases(t *testing.T) {
 	_, manager, cleanup := setupTestEnvironment(t)
 	defer cleanup()
@@ -708,211 +636,6 @@ func TestDeleteWorktreeByPath(t *testing.T) {
 	})
 }
 
-func TestCreateWorktreeWithGrenSymlink(t *testing.T) {
-	dir, manager, cleanup := setupTestEnvironment(t)
-	defer cleanup()
-
-	ctx := context.Background()
-
-	t.Run("creates symlink to .gren directory", func(t *testing.T) {
-		// The setup already creates .gren directory
-		grenDir := filepath.Join(dir, ".gren")
-		if _, err := os.Stat(grenDir); os.IsNotExist(err) {
-			t.Skip(".gren directory not found in test setup")
-		}
-
-		// Use absolute worktree dir so symlink path resolution works correctly
-		worktreeDir := filepath.Join(filepath.Dir(dir), filepath.Base(dir)+"-symlink-worktrees")
-		os.MkdirAll(worktreeDir, 0755)
-		defer os.RemoveAll(worktreeDir)
-
-		req := CreateWorktreeRequest{
-			Name:        "symlink-test",
-			IsNewBranch: true,
-			WorktreeDir: worktreeDir,
-		}
-
-		_, err := manager.CreateWorktree(ctx, req)
-		if err != nil {
-			t.Fatalf("CreateWorktree() error: %v", err)
-		}
-
-		// Check if .gren symlink exists in the worktree
-		worktreePath := filepath.Join(worktreeDir, "symlink-test")
-		symlinkPath := filepath.Join(worktreePath, ".gren")
-		info, err := os.Lstat(symlinkPath)
-		if err != nil {
-			// Symlink creation is best-effort, log but don't fail if it doesn't exist
-			t.Logf(".gren symlink not created (may be expected depending on path resolution): %v", err)
-			return
-		}
-
-		if info.Mode()&os.ModeSymlink == 0 {
-			t.Error(".gren is not a symlink")
-		}
-	})
-}
-
-func TestCreateWorktreeWithPostCreateHook(t *testing.T) {
-	dir, manager, cleanup := setupTestEnvironment(t)
-	defer cleanup()
-
-	ctx := context.Background()
-
-	t.Run("runs post-create hook when configured", func(t *testing.T) {
-		// Create a post-create hook script
-		hookScript := `#!/bin/sh
-echo "Hook executed" > "$1/.hook-executed"
-`
-		hookPath := filepath.Join(dir, ".gren", "post-create.sh")
-		if err := os.WriteFile(hookPath, []byte(hookScript), 0755); err != nil {
-			t.Fatalf("failed to create hook script: %v", err)
-		}
-
-		// Update config to use the hook
-		configPath := filepath.Join(dir, ".gren", "config.json")
-		configContent := fmt.Sprintf(`{
-			"main_worktree": %q,
-			"worktree_dir": "../test-worktrees",
-			"post_create_hook": ".gren/post-create.sh",
-			"version": "1.0.0"
-		}`, dir)
-		os.WriteFile(configPath, []byte(configContent), 0644)
-
-		req := CreateWorktreeRequest{
-			Name:        "hook-test",
-			IsNewBranch: true,
-		}
-
-		_, err := manager.CreateWorktree(ctx, req)
-		if err != nil {
-			t.Fatalf("CreateWorktree() error: %v", err)
-		}
-
-		// Find the worktree path
-		worktrees, _ := manager.ListWorktrees(ctx)
-		var worktreePath string
-		for _, wt := range worktrees {
-			if wt.Name == "hook-test" {
-				worktreePath = wt.Path
-				break
-			}
-		}
-
-		if worktreePath == "" {
-			t.Fatal("could not find worktree path")
-		}
-
-		// Check if hook created the marker file
-		markerPath := filepath.Join(worktreePath, ".hook-executed")
-		if _, err := os.Stat(markerPath); os.IsNotExist(err) {
-			t.Error("post-create hook was not executed (marker file not found)")
-		}
-	})
-
-	t.Run("handles missing hook gracefully", func(t *testing.T) {
-		// Update config to point to nonexistent hook
-		configPath := filepath.Join(dir, ".gren", "config.json")
-		configContent := fmt.Sprintf(`{
-			"main_worktree": %q,
-			"worktree_dir": "../test-worktrees",
-			"post_create_hook": ".gren/nonexistent-hook.sh",
-			"version": "1.0.0"
-		}`, dir)
-		os.WriteFile(configPath, []byte(configContent), 0644)
-
-		req := CreateWorktreeRequest{
-			Name:        "missing-hook-test",
-			IsNewBranch: true,
-		}
-
-		// Should not fail even with missing hook
-		_, err := manager.CreateWorktree(ctx, req)
-		if err != nil {
-			t.Fatalf("CreateWorktree() should not fail for missing hook: %v", err)
-		}
-	})
-}
-
-func TestCreateWorktreeWithRemoteBranch(t *testing.T) {
-	dir, manager, cleanup := setupTestEnvironment(t)
-	defer cleanup()
-
-	ctx := context.Background()
-
-	t.Run("creates worktree from remote tracking branch", func(t *testing.T) {
-		// Create a simulated remote by setting up origin
-		// First create a bare repo as "origin"
-		remoteDir, err := os.MkdirTemp("", "gren-remote-*")
-		if err != nil {
-			t.Fatalf("failed to create remote dir: %v", err)
-		}
-		defer os.RemoveAll(remoteDir)
-
-		// Init bare repo
-		exec.Command("git", "-C", remoteDir, "init", "--bare").Run()
-
-		// Add as remote
-		exec.Command("git", "-C", dir, "remote", "add", "origin", remoteDir).Run()
-
-		// Push current branch to remote
-		exec.Command("git", "-C", dir, "push", "-u", "origin", "HEAD").Run()
-
-		// Create a branch on remote only
-		// We do this by creating locally, pushing, and deleting locally
-		exec.Command("git", "-C", dir, "checkout", "-b", "remote-only-branch").Run()
-		exec.Command("git", "-C", dir, "push", "origin", "remote-only-branch").Run()
-
-		// Get back to main and delete local branch
-		currentBranch, _ := exec.Command("git", "-C", dir, "symbolic-ref", "--short", "HEAD").Output()
-		mainBranch := "master"
-		if len(currentBranch) > 0 {
-			mainBranch = string(currentBranch[:len(currentBranch)-1])
-		}
-
-		// Try to checkout main/master
-		if err := exec.Command("git", "-C", dir, "checkout", "master").Run(); err != nil {
-			exec.Command("git", "-C", dir, "checkout", "main").Run()
-			mainBranch = "main"
-		}
-
-		// Delete the local branch
-		exec.Command("git", "-C", dir, "branch", "-D", "remote-only-branch").Run()
-
-		// Fetch to make sure we have the remote ref
-		exec.Command("git", "-C", dir, "fetch", "origin").Run()
-
-		// Now create worktree from remote branch
-		req := CreateWorktreeRequest{
-			Name:        "from-remote",
-			Branch:      "remote-only-branch",
-			IsNewBranch: false,
-		}
-
-		_, err = manager.CreateWorktree(ctx, req)
-		if err != nil {
-			t.Logf("CreateWorktree from remote branch failed (may be expected in CI): %v", err)
-			// This test may fail in some CI environments without proper git setup
-			t.Skip("skipping remote branch test in environment without proper remote setup")
-		}
-
-		// Verify worktree was created
-		worktrees, _ := manager.ListWorktrees(ctx)
-		found := false
-		for _, wt := range worktrees {
-			if wt.Name == "from-remote" {
-				found = true
-				break
-			}
-		}
-		if !found {
-			t.Error("worktree 'from-remote' not found after creation from remote branch")
-		}
-
-		_ = mainBranch // used in checkout
-	})
-}
-
 func TestWorktreeInfoFields(t *testing.T) {
 	t.Run("WorktreeInfo struct has expected fields", func(t *testing.T) {
 		info := WorktreeInfo{
@@ -969,508 +692,6 @@ func TestCreateWorktreeRequestFields(t *testing.T) {
 		}
 		if req.WorktreeDir != "/custom/path" {
 			t.Errorf("WorktreeDir = %q, want '/custom/path'", req.WorktreeDir)
-		}
-	})
-}
-
-func TestIsBranchMerged(t *testing.T) {
-	dir, manager, cleanup := setupTestEnvironment(t)
-	defer cleanup()
-
-	t.Run("merged branch shows no unique commits after merge", func(t *testing.T) {
-		// Create a branch with a commit, then merge it
-		// After merging, commits are in main so hasUniqueCommits=false
-		exec.Command("git", "-C", dir, "checkout", "-b", "merged-branch").Run()
-		testFile := filepath.Join(dir, "merged-file.txt")
-		os.WriteFile(testFile, []byte("merged content"), 0644)
-		exec.Command("git", "-C", dir, "add", ".").Run()
-		exec.Command("git", "-C", dir, "commit", "-m", "Merged commit").Run()
-		exec.Command("git", "-C", dir, "checkout", "-").Run() // back to main/master
-		exec.Command("git", "-C", dir, "merge", "--no-ff", "merged-branch", "-m", "Merge merged-branch").Run()
-
-		merged, hasUniqueCommits := manager.isBranchMerged("merged-branch")
-		if !merged {
-			t.Error("isBranchMerged() merged=false for merged branch, want true")
-		}
-		// After merging, commits are now in main, so hasUniqueCommits is false
-		if hasUniqueCommits {
-			t.Error("isBranchMerged() hasUniqueCommits=true after merge, want false (commits are now in main)")
-		}
-	})
-
-	t.Run("empty branch returns merged=true, hasUniqueCommits=false", func(t *testing.T) {
-		// Create a branch with no unique commits
-		exec.Command("git", "-C", dir, "checkout", "-b", "empty-branch").Run()
-		exec.Command("git", "-C", dir, "checkout", "-").Run()
-
-		merged, hasUniqueCommits := manager.isBranchMerged("empty-branch")
-		if !merged {
-			t.Error("isBranchMerged() merged=false for empty branch, want true")
-		}
-		if hasUniqueCommits {
-			t.Error("isBranchMerged() hasUniqueCommits=true for empty branch, want false")
-		}
-	})
-
-	t.Run("unmerged branch returns false", func(t *testing.T) {
-		// Create a branch with a new commit (not merged)
-		exec.Command("git", "-C", dir, "checkout", "-b", "unmerged-branch").Run()
-		testFile := filepath.Join(dir, "unmerged-file.txt")
-		os.WriteFile(testFile, []byte("unmerged content"), 0644)
-		exec.Command("git", "-C", dir, "add", ".").Run()
-		exec.Command("git", "-C", dir, "commit", "-m", "Unmerged commit").Run()
-		exec.Command("git", "-C", dir, "checkout", "-").Run()
-
-		merged, _ := manager.isBranchMerged("unmerged-branch")
-		if merged {
-			t.Error("isBranchMerged() merged=true for unmerged branch, want false")
-		}
-	})
-
-	t.Run("nonexistent branch returns false", func(t *testing.T) {
-		merged, _ := manager.isBranchMerged("nonexistent-branch-xyz")
-		if merged {
-			t.Error("isBranchMerged() merged=true for nonexistent branch, want false")
-		}
-	})
-}
-
-func TestIsRemoteBranchGone(t *testing.T) {
-	dir, manager, cleanup := setupTestEnvironment(t)
-	defer cleanup()
-
-	t.Run("branch without tracking returns false", func(t *testing.T) {
-		// Create a local branch without remote tracking
-		exec.Command("git", "-C", dir, "checkout", "-b", "local-only-branch").Run()
-		exec.Command("git", "-C", dir, "checkout", "-").Run()
-
-		result := manager.isRemoteBranchGone("local-only-branch")
-		if result {
-			t.Error("isRemoteBranchGone() = true for local-only branch, want false")
-		}
-	})
-
-	t.Run("branch with active remote returns false", func(t *testing.T) {
-		// Set up a remote
-		remoteDir, err := os.MkdirTemp("", "gren-remote-test-*")
-		if err != nil {
-			t.Fatalf("failed to create remote dir: %v", err)
-		}
-		defer os.RemoveAll(remoteDir)
-
-		exec.Command("git", "-C", remoteDir, "init", "--bare").Run()
-		exec.Command("git", "-C", dir, "remote", "add", "origin", remoteDir).Run()
-
-		// Create branch and push
-		exec.Command("git", "-C", dir, "checkout", "-b", "remote-active").Run()
-		exec.Command("git", "-C", dir, "push", "-u", "origin", "remote-active").Run()
-		exec.Command("git", "-C", dir, "checkout", "-").Run()
-
-		result := manager.isRemoteBranchGone("remote-active")
-		if result {
-			t.Error("isRemoteBranchGone() = true for branch with active remote, want false")
-		}
-	})
-
-	t.Run("branch with gone remote returns true", func(t *testing.T) {
-		// Set up a remote if not already done
-		remoteDir, err := os.MkdirTemp("", "gren-remote-gone-*")
-		if err != nil {
-			t.Fatalf("failed to create remote dir: %v", err)
-		}
-		defer os.RemoveAll(remoteDir)
-
-		// Remove existing origin if any and add new one
-		exec.Command("git", "-C", dir, "remote", "remove", "origin").Run()
-		exec.Command("git", "-C", remoteDir, "init", "--bare").Run()
-		exec.Command("git", "-C", dir, "remote", "add", "origin", remoteDir).Run()
-
-		// Create branch, push, then delete remote branch
-		exec.Command("git", "-C", dir, "checkout", "-b", "remote-gone-branch").Run()
-		exec.Command("git", "-C", dir, "push", "-u", "origin", "remote-gone-branch").Run()
-		exec.Command("git", "-C", dir, "checkout", "-").Run()
-
-		// Delete the remote branch
-		exec.Command("git", "-C", dir, "push", "origin", "--delete", "remote-gone-branch").Run()
-
-		// Fetch to update tracking info
-		exec.Command("git", "-C", dir, "fetch", "--prune").Run()
-
-		result := manager.isRemoteBranchGone("remote-gone-branch")
-		if !result {
-			t.Error("isRemoteBranchGone() = false for branch with gone remote, want true")
-		}
-	})
-}
-
-func TestEnrichStaleStatus(t *testing.T) {
-	dir, manager, cleanup := setupTestEnvironment(t)
-	defer cleanup()
-
-	t.Run("main worktree stays active", func(t *testing.T) {
-		wt := &WorktreeInfo{
-			Name:   "main",
-			Branch: "main",
-			IsMain: true,
-			Status: "clean",
-		}
-
-		manager.enrichStaleStatus(wt)
-
-		if wt.BranchStatus != "active" {
-			t.Errorf("BranchStatus = %q, want 'active'", wt.BranchStatus)
-		}
-		if wt.StaleReason != "" {
-			t.Errorf("StaleReason = %q, want empty", wt.StaleReason)
-		}
-	})
-
-	t.Run("missing worktree stays active", func(t *testing.T) {
-		wt := &WorktreeInfo{
-			Name:   "missing-wt",
-			Branch: "some-branch",
-			IsMain: false,
-			Status: "missing",
-		}
-
-		manager.enrichStaleStatus(wt)
-
-		if wt.BranchStatus != "active" {
-			t.Errorf("BranchStatus = %q, want 'active'", wt.BranchStatus)
-		}
-	})
-
-	t.Run("detached HEAD stays active", func(t *testing.T) {
-		wt := &WorktreeInfo{
-			Name:   "detached-wt",
-			Branch: "(detached)",
-			IsMain: false,
-			Status: "clean",
-		}
-
-		manager.enrichStaleStatus(wt)
-
-		if wt.BranchStatus != "active" {
-			t.Errorf("BranchStatus = %q, want 'active'", wt.BranchStatus)
-		}
-	})
-
-	t.Run("bare repo stays active", func(t *testing.T) {
-		wt := &WorktreeInfo{
-			Name:   "bare-wt",
-			Branch: "(bare)",
-			IsMain: false,
-			Status: "clean",
-		}
-
-		manager.enrichStaleStatus(wt)
-
-		if wt.BranchStatus != "active" {
-			t.Errorf("BranchStatus = %q, want 'active'", wt.BranchStatus)
-		}
-	})
-
-	t.Run("branch with no unique commits becomes stale", func(t *testing.T) {
-		// Create an empty branch (no commits) - shows as stale with no_unique_commits
-		exec.Command("git", "-C", dir, "checkout", "-b", "stale-empty").Run()
-		exec.Command("git", "-C", dir, "checkout", "-").Run()
-
-		wt := &WorktreeInfo{
-			Name:   "stale-empty",
-			Branch: "stale-empty",
-			IsMain: false,
-			Status: "clean",
-		}
-
-		manager.enrichStaleStatus(wt)
-
-		if wt.BranchStatus != "stale" {
-			t.Errorf("BranchStatus = %q, want 'stale'", wt.BranchStatus)
-		}
-		// Empty branches show as "no_unique_commits"
-		if wt.StaleReason != "no_unique_commits" {
-			t.Errorf("StaleReason = %q, want 'no_unique_commits'", wt.StaleReason)
-		}
-	})
-
-	t.Run("active branch stays active", func(t *testing.T) {
-		// Create a branch with unmerged commits
-		exec.Command("git", "-C", dir, "checkout", "-b", "active-branch").Run()
-		testFile := filepath.Join(dir, "active-file.txt")
-		os.WriteFile(testFile, []byte("active content"), 0644)
-		exec.Command("git", "-C", dir, "add", ".").Run()
-		exec.Command("git", "-C", dir, "commit", "-m", "Active commit").Run()
-		exec.Command("git", "-C", dir, "checkout", "-").Run()
-
-		wt := &WorktreeInfo{
-			Name:   "active-branch",
-			Branch: "active-branch",
-			IsMain: false,
-			Status: "clean",
-		}
-
-		manager.enrichStaleStatus(wt)
-
-		if wt.BranchStatus != "active" {
-			t.Errorf("BranchStatus = %q, want 'active'", wt.BranchStatus)
-		}
-		if wt.StaleReason != "" {
-			t.Errorf("StaleReason = %q, want empty", wt.StaleReason)
-		}
-	})
-}
-
-func TestStaleStatusInListWorktrees(t *testing.T) {
-	dir, manager, cleanup := setupTestEnvironment(t)
-	defer cleanup()
-
-	ctx := context.Background()
-
-	t.Run("ListWorktrees includes stale status", func(t *testing.T) {
-		// Create a branch with a commit, then merge it
-		exec.Command("git", "-C", dir, "checkout", "-b", "stale-in-list").Run()
-		// Add a commit so there's something to merge
-		testFile := filepath.Join(dir, "stale-list-file.txt")
-		os.WriteFile(testFile, []byte("stale list content"), 0644)
-		exec.Command("git", "-C", dir, "add", ".").Run()
-		exec.Command("git", "-C", dir, "commit", "-m", "Stale list commit").Run()
-		exec.Command("git", "-C", dir, "checkout", "-").Run()
-		exec.Command("git", "-C", dir, "merge", "--no-ff", "stale-in-list", "-m", "Merge stale-in-list").Run()
-
-		// Create a worktree for the merged branch
-		worktreeDir := filepath.Join(filepath.Dir(dir), "test-worktrees")
-		os.MkdirAll(worktreeDir, 0755)
-		worktreePath := filepath.Join(worktreeDir, "stale-in-list")
-		exec.Command("git", "-C", dir, "worktree", "add", worktreePath, "stale-in-list").Run()
-
-		worktrees, err := manager.ListWorktrees(ctx)
-		if err != nil {
-			t.Fatalf("ListWorktrees() error: %v", err)
-		}
-
-		// Find the stale worktree
-		var staleWt *WorktreeInfo
-		for i := range worktrees {
-			if worktrees[i].Branch == "stale-in-list" {
-				staleWt = &worktrees[i]
-				break
-			}
-		}
-
-		if staleWt == nil {
-			t.Fatal("stale-in-list worktree not found")
-		}
-
-		if staleWt.BranchStatus != "stale" {
-			t.Errorf("BranchStatus = %q, want 'stale'", staleWt.BranchStatus)
-		}
-		// After merging, commits are in main, so shows as "no_unique_commits"
-		if staleWt.StaleReason != "no_unique_commits" {
-			t.Errorf("StaleReason = %q, want 'no_unique_commits'", staleWt.StaleReason)
-		}
-	})
-}
-
-// TestSubmoduleDetection tests that worktrees with submodules are detected
-func TestSubmoduleDetection(t *testing.T) {
-	dir, manager, cleanup := setupTestEnvironment(t)
-	defer cleanup()
-
-	ctx := context.Background()
-
-	t.Run("detect HasSubmodules when .gitmodules exists", func(t *testing.T) {
-		// Create a worktree
-		req := CreateWorktreeRequest{
-			Name:        "submodule-test",
-			IsNewBranch: true,
-		}
-		_, err := manager.CreateWorktree(ctx, req)
-		if err != nil {
-			t.Fatalf("failed to create worktree: %v", err)
-		}
-
-		// Find the worktree path
-		worktrees, _ := manager.ListWorktrees(ctx)
-		var worktreePath string
-		for _, wt := range worktrees {
-			if wt.Name == "submodule-test" {
-				worktreePath = wt.Path
-				// Before adding .gitmodules, should not have submodules
-				if wt.HasSubmodules {
-					t.Error("HasSubmodules should be false before adding .gitmodules")
-				}
-				break
-			}
-		}
-
-		if worktreePath == "" {
-			t.Fatal("could not find worktree path")
-		}
-
-		// Create a .gitmodules file to simulate submodule
-		gitmodulesPath := filepath.Join(worktreePath, ".gitmodules")
-		err = os.WriteFile(gitmodulesPath, []byte(`[submodule "test-submodule"]
-	path = test-submodule
-	url = https://github.com/example/test.git
-`), 0644)
-		if err != nil {
-			t.Fatalf("failed to create .gitmodules: %v", err)
-		}
-
-		// Re-list worktrees to check detection
-		worktrees, err = manager.ListWorktrees(ctx)
-		if err != nil {
-			t.Fatalf("ListWorktrees() error: %v", err)
-		}
-
-		// Find the worktree and check HasSubmodules
-		var foundWt *WorktreeInfo
-		for i := range worktrees {
-			if worktrees[i].Name == "submodule-test" {
-				foundWt = &worktrees[i]
-				break
-			}
-		}
-
-		if foundWt == nil {
-			t.Fatal("worktree not found after re-listing")
-		}
-
-		if !foundWt.HasSubmodules {
-			t.Error("HasSubmodules should be true after adding .gitmodules")
-		}
-
-		// Clean up
-		manager.DeleteWorktree(ctx, "submodule-test", true)
-	})
-
-	t.Run("HasSubmodules false when no .gitmodules", func(t *testing.T) {
-		// Create a worktree without submodules
-		req := CreateWorktreeRequest{
-			Name:        "no-submodule-test",
-			IsNewBranch: true,
-		}
-		_, err := manager.CreateWorktree(ctx, req)
-		if err != nil {
-			t.Fatalf("failed to create worktree: %v", err)
-		}
-
-		worktrees, _ := manager.ListWorktrees(ctx)
-		for _, wt := range worktrees {
-			if wt.Name == "no-submodule-test" {
-				if wt.HasSubmodules {
-					t.Error("HasSubmodules should be false for worktree without .gitmodules")
-				}
-				break
-			}
-		}
-
-		// Clean up
-		manager.DeleteWorktree(ctx, "no-submodule-test", true)
-	})
-
-	// Test worktree with submodule detection and deletion (mock .gitmodules file)
-	t.Run("delete worktree with submodule config", func(t *testing.T) {
-		// Create a worktree
-		req := CreateWorktreeRequest{
-			Name:        "submodule-test",
-			IsNewBranch: true,
-		}
-		_, err := manager.CreateWorktree(ctx, req)
-		if err != nil {
-			t.Fatalf("failed to create worktree: %v", err)
-		}
-
-		// Find the worktree path
-		worktrees, _ := manager.ListWorktrees(ctx)
-		var worktreePath string
-		for _, wt := range worktrees {
-			if wt.Name == "submodule-test" {
-				worktreePath = wt.Path
-				break
-			}
-		}
-
-		if worktreePath == "" {
-			t.Fatal("could not find worktree path")
-		}
-
-		// Create a mock .gitmodules file (simulates submodule presence)
-		gitmodulesPath := filepath.Join(worktreePath, ".gitmodules")
-		err = os.WriteFile(gitmodulesPath, []byte(`[submodule "test-submodule"]
-	path = test-submodule
-	url = https://github.com/example/test.git
-`), 0644)
-		if err != nil {
-			t.Fatalf("failed to create .gitmodules: %v", err)
-		}
-
-		// Re-list and verify detection
-		worktrees, _ = manager.ListWorktrees(ctx)
-		var foundWt *WorktreeInfo
-		for i := range worktrees {
-			if worktrees[i].Name == "submodule-test" {
-				foundWt = &worktrees[i]
-				break
-			}
-		}
-
-		if foundWt == nil {
-			t.Fatal("worktree not found")
-		}
-
-		if !foundWt.HasSubmodules {
-			t.Error("HasSubmodules should be true for worktree with .gitmodules")
-		}
-
-		// Delete the worktree with force flag (needed for submodules)
-		err = manager.DeleteWorktree(ctx, "submodule-test", true)
-		if err != nil {
-			t.Errorf("DeleteWorktree() should succeed for worktree with submodules: %v", err)
-		}
-
-		// Verify it's gone
-		worktrees, _ = manager.ListWorktrees(ctx)
-		for _, wt := range worktrees {
-			if wt.Name == "submodule-test" {
-				t.Error("worktree still exists after deletion")
-			}
-		}
-	})
-
-	// Test in main worktree to ensure it handles submodules there too
-	t.Run("main worktree with submodule detected", func(t *testing.T) {
-		// Add .gitmodules to main worktree
-		gitmodulesPath := filepath.Join(dir, ".gitmodules")
-		err := os.WriteFile(gitmodulesPath, []byte(`[submodule "main-submodule"]
-	path = main-submodule
-	url = https://github.com/example/test.git
-`), 0644)
-		if err != nil {
-			t.Fatalf("failed to create .gitmodules in main: %v", err)
-		}
-		defer os.Remove(gitmodulesPath)
-
-		worktrees, err := manager.ListWorktrees(ctx)
-		if err != nil {
-			t.Fatalf("ListWorktrees() error: %v", err)
-		}
-
-		// Find the main worktree
-		var mainWt *WorktreeInfo
-		for i := range worktrees {
-			if worktrees[i].IsMain {
-				mainWt = &worktrees[i]
-				break
-			}
-		}
-
-		if mainWt == nil {
-			t.Fatal("main worktree not found")
-		}
-
-		if !mainWt.HasSubmodules {
-			t.Error("HasSubmodules should be true for main worktree with .gitmodules")
 		}
 	})
 }
